jsとajaxを使っていいね！機能追加

+++ app/controllers/likes_controller.rb
@@ -1,13 +1,15 @@
 class LikesController < ApplicationController
 
   def create
-    @like = Like.create(user_id: current_user.id, photopost_id: params[:photopost_id])
-    @likes = Like.where(photopost_id: params[:photopost_id])
+    @photopost = Photopost.find(params[:photopost_id])
+    @like = current_user.likes.create(photopost_id: params[:photopost_id])
+    render 'likes/create'
   end
 
   def destroy
-    like = Likes.find_by(user_id: current_user.id, photopost_id: params[:photopost_id])
-    like.destroy
-    @likes = Like.where(photopost_id: params[:photopost_id])
+    @photopost = Photopost.find(params[:photopost_id])
+    like = current_user.likes.find_by(photopost_id: params[:photopost_id])
+    like.destroy if like
+    render 'likes/create'
   end
 end
